moving to non-generic token retriever
