+++ access-token-management/src/AccessTokenManagement/ITokenRetriever.cs
@@ -6,10 +6,10 @@ using Duende.AccessTokenManagement.OTel;
 namespace Duende.AccessTokenManagement;
 
 /// <summary>
-/// Interface for retrieving access tokens, on behalf of <see cref="AccessTokenHandler{TTokenRetriever,TToken}"/>
+/// Interface for retrieving access tokens, on behalf of <see cref="AccessTokenRequestHandler"/>
 /// </summary>
-/// <typeparam name="T">The type of token to retrieve</typeparam>
-public interface ITokenRetriever<T> where T : ClientCredentialsToken
+
+public interface ITokenRetriever
 {
     /// <summary>
     /// Method that retrieves the actual access token
@@ -18,7 +18,7 @@ public interface ITokenRetriever<T> where T : ClientCredentialsToken
     /// <param name="forceTokenRefresh"></param>
     /// <param name="cancellationToken"></param>
     /// <returns></returns>
-    Task<T> GetToken(HttpRequestMessage request, bool forceTokenRefresh, CancellationToken cancellationToken);
+    Task<ClientCredentialsToken> GetToken(HttpRequestMessage request, bool forceTokenRefresh, CancellationToken cancellationToken);
 
     /// <summary>
     /// The type of token that's requested, for metrics
